Attempt to grok https://github.com/ingydotnet/perl5-pm/issues/1
I refactored the tests to separate exporter form subexporter.

exporter can be adjusted with a global variable, so I got the test passing, but it's prbably wrong.

I think that lexwrap is messing up the call stack somehow.

I'd rather pull the lexwrap code directly into perl5 and not dep on it.